move mock to separate rust module
