+++ cosmwasm/contracts/yield-ward/src/tests/mocks/mod.rs
@@ -1,0 +1,1 @@
+pub mod ibc_module_mock;
