+++ cosmwasm/contracts/yield-ward/src/tests/mod.rs
@@ -1,6 +1,7 @@
 mod encoding;
 mod instantiate;
 mod lpt;
+mod mocks;
 mod reinit;
 mod stake;
 mod unstake;
